requirements.txt
pandas
numpy
scikit-learn
xgboost
joblib
python-dotenv
openai
matplotlib
